Move methods from TurnView to game_logic
